requirements.txt
# Core dependencies
python-dotenv>=1.0.0
datasets>=2.14.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
streamlit>=1.28.0

# LangChain
langchain>=0.0.350
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.0.0

# Optional: For local LLM support (uncomment if needed)
# langchain-community>=0.0.10
# ollama>=0.1.0

# Development (optional)
# pytest>=7.4.0
# black>=23.9.0
# flake8>=6.1.0

